% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{argChecks}
\alias{argChecks}
\title{Defensive programming for data re-ordering}
\usage{
argChecks(Y, W, X, N = NULL)
}
\arguments{
\item{Y}{A response vector (1 x n)}

\item{W}{A treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{X}{A covariate matrix (p x n) with no intercept. When
match_on="existing", then X must be a vector (1 x n) of user-specified
propensity scores.}

\item{N}{The number of unique units}
}
\description{
Defensive programming for data re-ordering
}

