% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchImputePO.R
\name{averageMultipleMatches}
\alias{averageMultipleMatches}
\title{Plumbing function for one-to-many matches}
\usage{
averageMultipleMatches(num_matches, matched_outcomes, orig_outcomes = NULL)
}
\arguments{
\item{num_matches}{Either \code{M_matches} or \code{J_var_matches}}

\item{matched_outcomes}{These are the \code{num_matches}-many imputed
potential outcomes. These need to be averaged if \code{num_matches}>1.}

\item{orig_outcomes}{When called from \code{\link{calcSigSqAI2006}}, these
are repeated outcomes that simply need to be subsetted; in the case of
\code{\link{wrangleImputations}} this is left to \code{NULL} and is
ignored.}
}
\value{
A list including the averaged \code{matched_outcomes}, and also
  \code{orig_outcomes} which is sometimes \code{NULL}.
}
\description{
This is called from \code{\link{wrangleImputations}} and from
\code{\link{calcSigSqAI2006}}.
}
