% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSigSq.R
\name{calcSigSqAI2006}
\alias{calcSigSqAI2006}
\title{Calculates \code{sigsqiw}}
\usage{
calcSigSqAI2006(match_output, J_var_matches)
}
\arguments{
\item{match_output}{Output of the \code{Matching::Match()} function for the
same-treatment matching, from \code{\link{estSigSq}}.}

\item{J_var_matches}{Number of matches when estimating sigmasq(X,W) as in
Abadie and Imbens 2006. Currently can only support J=1.}
}
\value{
A vector of the \code{sigsqiw} values for those individuals observed
  to have the \code{kk}th treatment level
}
\description{
This function estimates the conditional variance as seen in equation 14 in
Section 4.1 of Abadie and Imbens 2006 Econometrica. The matching procedure in
\code{\link{estSigSq}} matches units within the same treatment level (level
kk), and compares their outcomes to estimate a variance component. This
function was introduced to extend this variance matching procedure to
\code{J_var_matches >=1}, as it takes care of some of the bookeeping aspects
of one-to-many matching.
}
