% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{determineIDs}
\alias{determineIDs}
\title{Determines Unique Unit Identifiers}
\usage{
determineIDs(Y, W, X)
}
\arguments{
\item{Y}{A response vector (1 x n)}

\item{W}{A treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{X}{A covariate matrix (p x n) with no intercept. When
match_on="existing", then X must be a vector (1 x n) of user-specified
propensity scores.}
}
\value{
\code{unit_ids}
}
\description{
This function attempts to determine unique identifying information,
\code{unit_ids}, for each unit in the dataset. Users can apply this function
on their raw data ahead of using \code{\link{multiMatch}} to ensure that the
matching procedure will work. \code{unit_ids} will be used to identify study
units in some of the information output from \code{\link{multiMatch}}.
}

