% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau_legacy.R
\name{estimateTau_legacy}
\alias{estimateTau_legacy}
\title{Estimate the tau contrasts, perhaps with variance estimates}
\usage{
estimateTau_legacy(trtlevels, meanw, trtnumber, taunumber, N, Yiw = NULL,
  Kiw = NULL, sigsqiw = NULL, W = NULL, ...)
}
\arguments{
\item{trtlevels}{vector of the unique levels of treatment W}

\item{meanw}{vector of the estimated mean w.r.t. each treatment w}

\item{trtnumber}{a scalar, the number of treatment levels}

\item{taunumber}{a scalar, the number of tau contrasts to estimate

  To estimate variance in Matching using AI2006 method in MatchX and
  MatchGPS, all of the following must be supplied. Otherwise, they default to
  NULL. \itemize{

  \item @param Yiw

  \item @param Kiw

  \item @param sigsqiw

  \item @param W }}

\item{N}{A scalar for the number of rows in the data}

\item{Yiw}{Matrix of all imputed potential outcomes}

\item{Kiw}{Vector of times each unit is matched to}

\item{sigsqiw}{Estimated sigma squared, from AI2006}

\item{W}{a treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{...}{the dots argument}
}
\value{
a list including the dataframes for estimates for tau and for mu
}
\description{
This is a major plumbing function for the package. It is "legacy" and
designed to work with the following functions:
\code{\link{multilevelGPSMatch}}, \code{\link{multilevelMatchX}}, and
\code{\link{multilevelGPSStratification}}. The \code{\link{estimateTau}}
function is designed to work with \code{\link{multiMatch}} which is the
future direction of the matching estimators.
}
\seealso{
\code{\link{estimateTau}}; \code{\link{multilevelGPSMatch}};
  \code{\link{multilevelMatchX}}; \code{\link{multilevelGPSStratification}}
}
