% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTrtModel.R
\name{estimateTrtModel}
\alias{estimateTrtModel}
\title{Estimate Treatment Model for Generalized Propensity Scores}
\usage{
estimateTrtModel(W, X, match_on, model_options, ...)
}
\arguments{
\item{W}{A treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{X}{A covariate matrix (p x n) with no intercept. When
match_on="existing", then X must be a vector (1 x n) of user-specified
propensity scores.}

\item{match_on}{"multinom", "polr", "existing", or "covariates",}

\item{model_options}{A list of the options to pass to propensity model.
Currently under development. Can only pass reference level to multinomial
logisitc regression.}

\item{...}{the dots argument}
}
\description{
Note that the model_options argument must be a list with reference_level
element
}

