% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estSigSq.R
\name{estSigSq}
\alias{estSigSq}
\title{Perform matching methods to estimate variance for one treatment level}
\usage{
estSigSq(X, Y, which_same_trt, N_this_trt, var_options, match_on,
  var_args_AI2016)
}
\arguments{
\item{X}{A covariate matrix (p x n) with no intercept. When
match_on="existing", then X must be a vector (1 x n) of user-specified
propensity scores.}

\item{Y}{A response vector (1 x n)}

\item{which_same_trt}{Numeric vector providing indices of individuals
observed to have treatment level kk}

\item{N_this_trt}{The number of units observed to have this treatment level.}

\item{var_options}{Options for carrying out matching for variance estimation.}

\item{match_on}{"multinom", "polr", "existing", or "covariates",}

\item{var_args_AI2016}{A list of arguments for carrying out matching
  procedures to estimate components in the VarianceAI2016 variance estimates
  (see \code{\link{calcSigSqAI2016}}).

  Note that these variance components are combined in
  \code{\link{estimateTau}} (and perhaps \code{\link{calcSigSqAI2016}}).}
}
\value{
A list of one or two elements. \code{sigsqiw_kk} is a vector with the
  estimated variance component for units observed to have the \code{kk}^th treatment
  level. When \code{match_on = "multinom"}, the list will also have an
  element for \code{match_mat_AI2016_kk_two_cols}, which is a \code{Ntot}-by-2
  matrix with matching information that will eventually be passed to
  \code{\link{calcSigSqAI2016}}.
}
\description{
This function executes the matching methods to estimate the components for
estimating the variance (sigma squared) of the matching estimator. All
matching methods (specified with \code{match_on}) will estimate the variance
as in Abadie and Imbens (2006). When \code{match_on = "multinom"} this function
will carry out additional matching procedures to estimate the variance as
described in Abadie and Imbens (2016).
}
\references{
Abadie, A., & Imbens, G. W. (2006). Large sample properties of
  matching estimators for average treatment effects. econometrica, 74(1),
  235-267. \url{https://doi.org/10.1111/j.1468-0262.2006.00655.x}

  Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity
  score. Econometrica, 84(2), 781-807.
  \url{https://doi.org/10.3982/ECTA11293}
}
