% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{estVarAI2006}
\alias{estVarAI2006}
\title{Computes Estimated Asymptotic Variance of matching estimators.}
\usage{
estVarAI2006(N, W, M_matches, trt_level_1, trt_level_2, Yiw1, Yiw2, tau, Kiw,
  sigsqiw)
}
\arguments{
\item{N}{A scalar for the number of rows in the data}

\item{W}{A treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{M_matches}{Number of matches per unit for imputing potential outcomes,
as in Abadie and Imbens 2006. Currently can only support M=1.}

\item{trt_level_1}{Unique treatment level 1; aka \eqn{W_1} in \eqn{\tau(W_1,
W_2)}}

\item{trt_level_2}{Unique treatment level 2; aka \eqn{W_2} in \eqn{\tau(W_1,
W_2)}}

\item{Yiw1}{Vector of imputed outcomes for all units for \code{trt_level_2}.}

\item{Yiw2}{Vector of imputed outcomes for all units for \code{trt_level_2}.}

\item{tau}{Estimated value of \eqn{\tau(W_1, W_2)}.}

\item{Kiw}{A vector of times each unit is matched to}

\item{sigsqiw}{Estimated sigma squared (variance), from Abadie and Imbens
(2006)}
}
\value{
A single numeric value for the estimated asymptotic variance of the
  estimator.
}
\description{
See Theorem 7 of Abadie and Imbens (2006) Econometrica for the formula.
}
\references{
Abadie, A., & Imbens, G. W. (2006). Large sample properties of
  matching estimators for average treatment effects. econometrica, 74(1),
  235-267. \url{https://doi.org/10.1111/j.1468-0262.2006.00655.x}
}
