% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMatch.R
\name{estVarAI2012}
\alias{estVarAI2012}
\title{Estimate Variance as in Abadie & Imebens 2012 JASA}
\usage{
estVarAI2012(W, X_covars, Y, N, num_trts, trt_levels, match_mat_AI2012,
  prop_score_ests, vcov_coeff, tau_dfm, ...)
}
\arguments{
\item{W}{a treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{X_covars}{The matrix of covariates (not the matrix of estimated GPS)}

\item{Y}{a continuous response vector (1 x n)}

\item{N}{The total number of units}

\item{num_trts}{The number of unique treatments (3+), a scalar}

\item{trt_levels}{A vector (of length \code{num_trts} providing the unique
treatment levels}

\item{match_mat_AI2012}{Information on matches from
\code{\link{matchAllTreatments}}}

\item{prop_score_ests}{The matrix of propensity scores}

\item{vcov_coeff}{The variance-covariance (inverse fisher information) matrix
from the multinomial logistic regression model}

\item{tau_dfm}{The dataframe of estimates and estimated standard errors
output from \code{\link{estimateTau}}}

\item{...}{the dots argument}
}
\value{
A list with the updated \code{tau_dfm} includeing a column for
  \code{VarianceAI2012}, and a list object \code{AI2012_args} with
  potentially helpful extra information.
}
\description{
This function estimates variance of matching estimator when matching on
propensity scores that were estimated with multinomial logistic regression.
This method takes into account the uncertainty from the treatment model.
}
