% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getIDs}
\alias{getIDs}
\title{Get/grab identifiers from vector/matrix/dataframe}
\usage{
getIDs(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{names(x)}, \code{row.names(x)}, or \code{NULL}.
}
\description{
This is a helper function for \code{\link{determineIDs}}
}

