% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_plumbing.R
\name{matchAllTreatments}
\alias{matchAllTreatments}
\title{Estimate Imputed Outcomes and AI06 Variance Components}
\usage{
matchAllTreatments(N, X, W, Y, num_trts, trt_levels, N_per_trt, M_matches,
  J_var_matches, match_on, ...)
}
\arguments{
\item{N}{The total number of units}

\item{X}{A covariate matrix (p x n) with no intercept. When
match_on="existing", then X must be a vector (1 x n) of user-specified
propensity scores.}

\item{W}{a treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{Y}{a continuous response vector (1 x n)}

\item{num_trts}{The number of unique treatments (3+), a scalar}

\item{trt_levels}{A vector (of length \code{num_trts} providing the unique
treatment levels}

\item{N_per_trt}{A vector (of length \code{num_trts}) indicating the number
of units observed to have each treatment level}

\item{M_matches}{Number of matches per unit for imputing potential outcomes,
as in Abadie and Imbens 2006. Currently can only support M=1.}

\item{J_var_matches}{Number of matches when estimating sigmasq(X,W) as in
Abadie and Imbens 2006. Currently can only support J=1.}

\item{match_on}{"multinom", "polr", "existing", or "covariates",}

\item{...}{the dots argument}
}
\value{
A list including at most: \itemize{

  \item Yiw: A matrix of all imputed (or observed) potential outcomes for
  each unit \item mean_Yiw: A vector of the average (across all units) of
  estimated/imputed potential outcomes \item sigsqiw: Estimated variance
  components (from AI2006) for each unit \item Kiw: The vector of number of
  times unit i used as a match \item impute_match_data: extra information
  from the main matching procedure \item match_mat_AI2012: When
  \code{match_on=`multinom`} this additional information will be output for
  \code{\link{estVarAI2012}} }

  A few of the necessary arguments are output from the
  \code{\link{reorderByTreatment}} function.
}
\description{
Mostly calls a subfunction for each treatment level, which calls subfunctions
that either carry out the matched imputation or the variance estimation.
}
