% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelGPSStratification.r
\name{multilevelGPSStratification}
\alias{multilevelGPSStratification}
\title{Stratification on GPS with multilevel treatments}
\usage{
multilevelGPSStratification(Y, W, X, NS, GPSM = "multinomiallogisticReg",
  linearp = 0, nboot, model_options = list(reference_level = "1"))
}
\arguments{
\item{Y}{a continuous response vector (1 x n)}

\item{W}{a treatment vector (1 x n) with numerical values indicating treatment groups}

\item{X}{a covariate matrix (p x n) with no intercept}

\item{NS}{(only required in the function multilevelGPSStratification) the number of strata}

\item{GPSM}{an indicator of the methods used for estimating GPS, options include "multinomiallogisticReg", "ordinallogisticReg", and "existing"}

\item{linearp}{(only required in the function multilevelGPSStratification) an indicator of subclassification on GPS (=0) or linear predictor of GPS (=1)}

\item{nboot}{(only required in the function multilevelGPSStratification) the number of boot replicates for variance estimation}

\item{model_options}{A list of the options to pass to propensity model.
Currently under development. Can only pass reference level to multinomial
logisitc regression.}
}
\value{
according to \code{\link{estimateTau}}: a dataframe with two columns,
tauestimate, varestimate, where
tauestimate is a vector of estimates for pairwise treatment effects, and
varestimate is a vector of variance estimates, using bootstrapping method.
}
\description{
Stratification on GPS with multilevel treatments
}
\seealso{
\code{\link{multilevelGPSMatch}}; \code{\link{multilevelMatchX}}
}
