% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameContrast}
\alias{nameContrast}
\title{naming the matching contrasts}
\usage{
nameContrast(trt1, trt2)
}
\arguments{
\item{trt1}{former treatment level}

\item{trt2}{latter treatment level}
}
\description{
naming the matching contrasts
}
