% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameMu}
\alias{nameMu}
\title{Naming the matching population mean mu's}
\usage{
nameMu(trt)
}
\arguments{
\item{trt}{Treatment level}
}
\description{
Naming the matching population mean mu's
}
\examples{
 nameMu(1)

}
