% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.r
\name{overlap}
\alias{overlap}
\title{Function to determine overlap from Crump}
\usage{
overlap(PF.fit)
}
\arguments{
\item{PF.fit}{fitted propensity model}
}
\description{
Function to determine overlap from Crump
}
