% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{print.multiMatch}
\alias{print.multiMatch}
\title{Prints a summary of the estimates from a multiMatch object}
\usage{
\method{print}{multiMatch}(x, ...)
}
\arguments{
\item{x}{object of class "multiMatch"}

\item{...}{dots}
}
\description{
Prints a summary of the estimates from a multiMatch object
}
