% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{setIDs}
\alias{setIDs}
\title{Set ID's from vector/matrix/dataframe}
\usage{
setIDs(x, unit_ids)
}
\arguments{
\item{x}{An object}

\item{unit_ids}{Character vector to identify study units.}
}
\value{
names(x), row.names(x), or NULL.
}
\description{
This is a helper function for \code{\link{determineIDs}}
}
