% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3methods.R
\name{summary.multiMatch}
\alias{summary.multiMatch}
\title{Prints a summary of a multiMatch object}
\usage{
\method{summary}{multiMatch}(object, ...)
}
\arguments{
\item{object}{object of class "multiMatch"}

\item{...}{dots}
}
\description{
Prints a summary of a multiMatch object
}
\author{
Brian G. Barkley
}
