% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_output.R, R/tidying_utilities.R
\name{tidyOutput}
\alias{tidyOutput}
\alias{tidyOutput}
\title{DEPRECATED.Strcutres main output in a "tidy" datframe}
\usage{
tidyOutput(untidy_output)

tidyOutput(untidy_output)
}
\arguments{
\item{untidy_output}{A list of objects to output from
\code{\link{multilevelXMatch}}
\code{\link{multilevelGPSMatch}}
\code{\link{multilevelGPSStratification}}}

\item{untidy_output}{A list of objects to output from
\code{\link{multilevelXMatch}}
\code{\link{multilevelGPSMatch}}
\code{\link{multilevelGPSStratification}}}
}
\value{
a list containing
\itemize{

 \item results: A tidy dataframe continaing the following columns:
   \itemize{

    \item tauestimate:  a vector of estimates for pairwise treatment effects

    \item varestimate:  a vector of variance estimates for tauestimate,
    using Abadie & Imbens (2006)'s method

    \item varestimateAI2012:  a vector of variance estimates for
    tauestimate, when matching on the generalized propensity score, using
    Abadie & Imbens (2012)'s method.  This variance estimate takes into account
    of the uncertainty in estimating the GPS. This will be returned only for
    the \code{\link{multilevelGPSMatch}} function and fitting propensity
    scores with multinomial logistic regression.

   }
 \item analysis_idx: the index of units after trimming based on Crump et
 al. (2009)'s method. This will be returned only if trimming is used in
 \code{\link{multilevelGPSMatch}} or \code{\link{multilevelGPSStratification}}

}

a list containing
\itemize{

 \item results: A tidy dataframe continaing the following columns:
   \itemize{

    \item tauestimate:  a vector of estimates for pairwise treatment effects

    \item varestimate:  a vector of variance estimates for tauestimate,
    using Abadie & Imbens (2006)'s method

    \item varestimateAI2012:  a vector of variance estimates for
    tauestimate, when matching on the generalized propensity score, using
    Abadie & Imbens (2012)'s method.  This variance estimate takes into account
    of the uncertainty in estimating the GPS. This will be returned only for
    the \code{\link{multilevelGPSMatch}} function and fitting propensity
    scores with multinomial logistic regression.

   }
 \item analysisidx: the index of units after trimming based on Crump et
 al. (2009)'s method. This will be returned only if trimming is used in
 \code{\link{multilevelGPSMatch}} or \code{\link{multilevelGPSStratification}}

}
}
\description{
DEPRECATED.Strcutres main output in a "tidy" datframe

Strcutres main output in a "tidy" datframe
}
