% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{estimateTau}
\alias{estimateTau}
\title{Estimate the tau contrasts, perhaps with variance estimates}
\usage{
estimateTau(trtlevels, meanw, trtnumber, taunumber, N, Yiw = NULL,
  Kiw = NULL, sigsqiw = NULL, W = NULL)
}
\arguments{
\item{trtlevels}{vector of the unique levels of treatment W}

\item{meanw}{vector of the estimated mean w.r.t. each treatment w}

\item{trtnumber}{a scalar, the number of treatment levels}

\item{taunumber}{a scalar, the number of tau contrasts to estimate}

\item{N}{A scalar for the number of rows in the data
To estimate variance in Matching using AI2006 method in MatchX and MatchGPS,
all of the following must be supplied. Otherwise, they default to NULL.
  \itemize{

  \item @param Yiw:  a vector of estimates for pairwise treatment effects

  \item @param Kiw

  \item @param sigsqiw

  \item @param W
  }

 @seealso \code{\link{multilevelGPSMatch}}; \code{\link{multilevelMatchX}};
 \code{\link{multilevelGPSStratification}}}
}
\value{
a tidy dataframe including Estimate, Variance, and referencing the
  target parameter (tau)
}
\description{
This is a major plumbing function for the package
}
