% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau.R
\name{estimateTau}
\alias{estimateTau}
\title{Estimate the tau contrasts, perhaps with variance estimates}
\usage{
estimateTau(trt_levels, mean_Yiw, num_trts, num_contrasts, N, M_matches, Yiw,
  Kiw, sigsqiw, W, ...)
}
\arguments{
\item{trt_levels}{vector of the unique levels of treatment W}

\item{mean_Yiw}{vector of the estimated mean w.r.t. each treatment w}

\item{num_trts}{a scalar, the number of treatment levels}

\item{num_contrasts}{a scalar, the number of tau contrasts to estimate}

\item{N}{A scalar for the number of rows in the data}

\item{M_matches}{Number of matches per unit for imputing potential outcomes,
as in Abadie and Imbens 2006. Currently can only support M=1.}

\item{Yiw}{Matrix of all imputed potential outcomes}

\item{Kiw}{Vector of times each unit is matched to}

\item{sigsqiw}{Estimated sigma squared, from AI2006}

\item{W}{a treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{...}{the dots argument}
}
\value{
a list including the dataframes for estimates for tau and for mu
}
\description{
This is a major plumbing function for the package
}
\seealso{
\code{\link{multiMatch}};
}
