% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTau_legacy.R
\name{estimateTau_legacy}
\alias{estimateTau_legacy}
\title{Estimate the contrasts "tau", perhaps with variance estimates}
\usage{
estimateTau_legacy(trtlevels, meanw, trtnumber, taunumber, N, Yiw = NULL,
  Kiw = NULL, sigsqiw = NULL, W = NULL, ...)
}
\arguments{
\item{trtlevels}{A vector of the unique levels of treatment W}

\item{meanw}{A vector of the estimated mean w.r.t. each treatment w}

\item{trtnumber}{A scalar, the number of treatment levels}

\item{taunumber}{A scalar, the number of tau contrasts to estimate

  To estimate variance in Matching using AI2006 method in
  \code{\link{multilevelGPSMatch}} and \code{\link{multilevelMatchX}}, all of
  the following must be supplied. Otherwise, they default to NULL. \itemize{

  \item @param Yiw

  \item @param Kiw

  \item @param sigsqiw

  \item @param W }}

\item{N}{A scalar for the number of rows in the data}

\item{Yiw}{A matrix of all imputed potential outcomes}

\item{Kiw}{A vector of times each unit is matched to}

\item{sigsqiw}{Estimated sigma squared (variance), from Abadie and Imbens
(2006)}

\item{W}{A treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{...}{the dots argument}
}
\value{
a list including the dataframes for estimates for tau and for mu
}
\description{
This is a major plumbing function for the package. It is "legacy" and
designed to work with the following functions:
\code{\link{multilevelGPSMatch}}, \code{\link{multilevelMatchX}}, and
\code{\link{multilevelGPSStratification}}. The \code{\link{estimateTau}}
function is designed to work with \code{\link{multiMatch}} which is the
future direction of the matching estimators. Note that the AI2012 reference
is left for backwards-compatibility, although it references the Abadie and
Imbens (2016) Econometrica article.
}
\references{
Abadie, A., & Imbens, G. W. (2006). Large sample properties of
  matching estimators for average treatment effects. econometrica, 74(1),
  235-267. \url{https://doi.org/10.1111/j.1468-0262.2006.00655.x}

  Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity
  score. Econometrica, 84(2), 781-807.
  \url{https://doi.org/10.3982/ECTA11293}
}
\seealso{
\code{\link{estimateTau}}; \code{\link{multilevelGPSMatch}};
  \code{\link{multilevelMatchX}}; \code{\link{multilevelGPSStratification}}
}
