% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{getIDs}
\alias{getIDs}
\title{Get/grab ID's from vector/matrix/dataframe}
\usage{
getIDs(x)
}
\arguments{
\item{x}{An object}
}
\value{
names(x), row.names(x), or NULL.
}
\description{
This is a helper function for \code{\link{determineIDs}}
}
