% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchImputePO.R
\name{matchImputePO}
\alias{matchImputePO}
\title{Perform matching methods to estimate (or impute) potential outcomes for one
treatment level (level kk)}
\usage{
matchImputePO(X, Y, vec_diff_trt, which_diff_trt, vec_same_trt, which_same_trt,
  est_options)
}
\arguments{
\item{X}{A covariate matrix (p x n) with no intercept. When
match_on="existing", then X must be a vector (1 x n) of user-specified
propensity scores.}

\item{Y}{a continuous response vector (1 x n)}

\item{vec_diff_trt}{Boolean vector indicating which individuals observed to
have treatment level NOT equal to kk}

\item{which_diff_trt}{Numeric vector providing indices of individuals
observed to have treatment level  NOT equal to kk}

\item{vec_same_trt}{Boolean vector indicating which individuals observed to
have treatment level kk}

\item{which_same_trt}{Numeric vector providing indices of individuals
observed to have treatment level kk}

\item{est_options}{Options for carrying out matching for point estimation.}
}
\value{
A list, many of which are necessary arguments for
  \code{\link{estimateTau}}: \itemize{ \item \code{Yiw} The fully-imputed
  data set of all potential outcomes. \item \code{mean_Yiw} Mean of Yiw
  across all units i. These are the estimates of the population-level mean
  estimands. See \code{\link{estimateTau}}. \item \code{Kiw} The vector of
  number of times unit i used as a match in imputation. \item
  \code{impute_match_data} Extra information from the main matching
  procedure. }
}
\description{
This function executes the matching methods to estimate the components for
the point estimates of the population-level estimands. The components will be
assembled into the estimator in the function \code{\link{estimateTau}}.
}
