% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelGPSMatch.r
\name{multilevelGPSMatch}
\alias{multilevelGPSMatch}
\title{multilevel treatment matching on GPS}
\usage{
multilevelGPSMatch(Y, W, X, Trimming, GPSM = "multinomiallogisticReg")
}
\arguments{
\item{Y}{response vector (1 x n)}

\item{W}{treatment vector (1 x n)}

\item{X}{covariate matrix (p x n) with no intercept}
}
\value{
A list with 2 elements: tauestimate, varestimate
}
\description{
multilevel treatment matching on GPS
}

