% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelGPSMatch.r
\name{multilevelGPSMatch}
\alias{multilevelGPSMatch}
\title{Matching on GPS with multilevel treatments}
\usage{
multilevelGPSMatch(Y, W, X, Trimming, GPSM = "multinomiallogisticReg")
}
\arguments{
\item{Y}{a continuous response vector (1 x n)}

\item{W}{a treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{X}{a covariate matrix (p x n) with no intercept. When GPSM="existing",
then X must be a vector (1 x n) of user-specified propensity scores.}

\item{Trimming}{an indicator of whether trimming the sample to ensure overlap}

\item{GPSM}{an indicator of the methods used for estimating GPS, options
include "multinomiallogisticReg", "ordinallogisticReg" for proportional
odds or cumulative logit model, and "existing" for user-specified
propensity score via the parameter X.}
}
\value{
according to \code{\link{tidyOutput}}, including at most: \itemize{

  \item tauestimate:  a vector of estimates for pairwise treatment effects

  \item varestimate:  a vector of variance estimates for tauestimate, using
  Abadie&Imbens(2006)'s method

  \item varestimateAI2012:  a vector of variance estimates for tauestimate,
  when matching on the generalized propensity score, using
  Abadie&Imbens(2012)'s method. This variance estimate takes into account of
  the uncertainty in estimating the GPS.

  \item analysis_idx: a list containing the indices_kept (analyzed) and
  indices_dropped (trimmed) based on Crump et al. (2009)'s method.

  }
}
\description{
Matching on GPS with multilevel treatments
}
\examples{
  X<-c(5.5,10.6,3.1,8.7,5.1,10.2,9.8,4.4,4.9)
  Y<-c(102,105,120,130,100,80,94,108,96)
  W<-c(1,1,1,3,2,3,2,1,2)
  multilevelGPSMatch(Y,W,X,Trimming=0,GPSM="multinomiallogisticReg")
  multilevelGPSMatch(Y,W,X,Trimming=1,GPSM="multinomiallogisticReg")

}
\seealso{
\code{\link{multilevelMatchX}};
  \code{\link{multilevelGPSStratification}}
}
