% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelGPSStratification.r
\name{multilevelGPSStratification}
\alias{multilevelGPSStratification}
\title{Stratification on GPS with multilevel treatments}
\usage{
multilevelGPSStratification(Y, W, X, NS, GPSM = "multinomiallogisticReg",
  linearp = 0, nboot)
}
\arguments{
\item{Y}{A continuous response vector (1 x n)}

\item{W}{A treatment vector (1 x n) with numerical values indicating
treatment groups}

\item{X}{A covariate matrix (p x n) with no intercept}

\item{NS}{The number of strata: (only required in the function
\code{\link{multilevelGPSStratification}})}

\item{GPSM}{An indicator of the methods used for estimating GPS, options
include "multinomiallogisticReg", "ordinallogisticReg", and "existing"}

\item{linearp}{An indicator of subclassification on GPS (=0) or linear
predictor of GPS (=1): (only required in the function
\code{\link{multilevelGPSStratification}})}

\item{nboot}{The number of boot replicates for variance estimation: (only
required in the function \code{\link{multilevelGPSStratification}})}
}
\value{
A list with two elements,
  \code{tauestimate}, \code{varestimate}, where \code{tauestimate} is a
  vector of estimates for pairwise treatment effects, and \code{varestimate}
  is a vector of variance estimates, using bootstrapping method.
}
\description{
Stratification on GPS with multilevel treatments
}
\examples{

  set.seed(111)
  n    <- 5000*6
  # X1-X3 3 MVN var 2, 1, 1, covars 1, -1, -.5
  vars   <- c(2,1,1)
  covars <- c(1,-1,-.5)
  mu     <- c(0,0,0)
  tau    <- 1
  Sigma <- diag(vars)
  Sigma[2,1] <- Sigma[1,2] <- covars[1]
  Sigma[3,1] <- Sigma[1,3] <- covars[2]
  Sigma[3,2] <- Sigma[2,3] <- covars[3]
  trt1 <- 100; trt1
  trt2 <- 100; trt2
  trt3 <- 100; trt3
  # draw Xs
  X13 <- MASS::mvrnorm(n,mu=mu,Sigma=Sigma, empirical = FALSE)
  X1 <- X13[,1]
  X2 <- X13[,2]
  X3 <- X13[,3]
  X4 <- runif(n,-3,3)
  X5 <- rchisq(n, df=1)
  X6 <- rbinom(n,size=1,prob=.5)

  xb2 <- 0.1*(X1^2+X2+X3+X4+X5+X6)
  xb3 <- 0.1*(X1+X2^2+X3^2+X4+X5+X6)
  exb2<-exp(xb2)
  exb3<-exp(xb3)
  pi1<-1/(1+exp(xb2)+exp(xb3))
  pi2<-exp(xb2)/(1+exp(xb2)+exp(xb3))
  pi3<-exp(xb3)/(1+exp(xb2)+exp(xb3))
  pi<-cbind(pi1,pi2,pi3)
  apply(pi,2,mean)

  W<-matrix(NA,n,4)
  colnames(W)   <- c("W1","W2","W3","W")
  for(kk in 1:n){
    W[kk,1:3]<-stats::rmultinom(1, 1, prob = pi[kk,])
  }

  sim.dat <- data.frame(W,X1,X2,X3,X4,X5,X6)
  trt1.keep <- sample(which(sim.dat$W1==1),trt1,replace=FALSE)
  trt2.keep <- sample(which(sim.dat$W2==1),trt2,replace=FALSE)
  trt3.keep <- sample(which(sim.dat$W3==1),trt3,replace=FALSE)
  sim.dat <- sim.dat[c(trt1.keep,trt2.keep,trt3.keep),]
  sim.dat[,"W"]<-sim.dat[,"W1"]+2*sim.dat[,"W2"]+3*sim.dat[,"W3"]
  sim.dat[,"W"]<-as.factor(sim.dat[,"W"])
  W <- sim.dat[,"W"]
  X <- as.matrix(sim.dat[,names(sim.dat)[-c(1:4)]])
  X1 <- X[,"X1"]; X2 <- X[,"X2"]; X3 <- X[,"X3"]; X4 <- X[,"X4"]; X5 <- X[,"X5"];X6 <- X[,"X6"]

  # outcome: treatment effect is zero
  u  <- rnorm(nrow(X))
  # ouctome 1 (linear)
  Y <- 	(W==1)*(  X1 +   X2 +   X3 +   X4 +    X5-1 +     X6-0.5)+
    (W==2)*(2*X1 + 3*X2 +   X3 + 2*X4 + 2*(X5-1) + 2*(X6-0.5))+
    (W==3)*(3*X1 +   X2 + 2*X3 -   X4 -   (X5-1) -   (X6-0.5))+u

  match1<-multilevelMatchX(Y,W,X)
  match2<-multilevelGPSMatch(Y,W,X,Trimming=FALSE,GPSM="multinomiallogisticReg")
  match3<-multilevelGPSMatch(Y,W,X,Trimming=TRUE,GPSM="multinomiallogisticReg")
  match4<-multilevelGPSStratification(Y,W,X,NS=10,GPSM="multinomiallogisticReg",linearp=0,nboot=50)

  c(match1$tauestimate,match1$varestimate)
  c(match2$tauestimate,match2$varestimate)
  c(match3$tauestimate,match3$varestimate)
  c(match4$tauestimate,match4$varestimate)

}
\references{
Yang, S., Imbens G. W., Cui, Z., Faries, D. E., & Kadziola, Z.
  (2016) Propensity Score Matching and Subclassification in Observational
  Studies with Multi-Level Treatments. Biometrics, 72, 1055-1065.
  \url{https://doi.org/10.1111/biom.12505}

  Abadie, A., & Imbens, G. W. (2006). Large sample properties of matching
  estimators for average treatment effects. econometrica, 74(1), 235-267.
  \url{https://doi.org/10.1111/j.1468-0262.2006.00655.x}

  Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity
  score. Econometrica, 84(2), 781-807.
  \url{https://doi.org/10.3982/ECTA11293}
}
\seealso{
\code{\link{multilevelGPSMatch}}; \code{\link{multilevelMatchX}}
}

