% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelGPSStratification.r
\name{multilevelGPSStratification}
\alias{multilevelGPSStratification}
\title{multilevel treatment stratification on GPS}
\usage{
multilevelGPSStratification(Y, W, X, NS, GPSM = "multinomiallogisticReg",
  linearp = 0, nboot)
}
\arguments{
\item{Y}{response vector (1 x n)}

\item{W}{treatment vector (1 x n)}

\item{X}{covariate matrix (p x n) with no intercept}
}
\value{
A list with 2 elements: tauestimate, varestimate
}
\description{
multilevel treatment stratification on GPS
}

