% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelGPSStratification.r
\name{multilevelGPSStratification}
\alias{multilevelGPSStratification}
\title{Stratification on GPS with multilevel treatments}
\usage{
multilevelGPSStratification(Y, W, X, NS, GPSM = "multinomiallogisticReg",
  linearp = 0, nboot, model_options = list(reference_level = "1"))
}
\arguments{
\item{Y}{A continuous response vector (1 x n)}

\item{W}{A treatment vector (1 x n) with numerical values indicating treatment groups}

\item{X}{A covariate matrix (p x n) with no intercept}

\item{NS}{(only required in the function
\code{\link{multilevelGPSStratification}}) the number of strata}

\item{GPSM}{an indicator of the methods used for estimating GPS, options
include "multinomiallogisticReg", "ordinallogisticReg", and "existing"}

\item{linearp}{(only required in the function
\code{\link{multilevelGPSStratification}}) an indicator of
subclassification on GPS (=0) or linear predictor of GPS (=1)}

\item{nboot}{(only required in the function
\code{\link{multilevelGPSStratification}}) the number of boot replicates
for variance estimation}

\item{model_options}{A list of the options to pass to propensity model.
Currently under development. Can only pass reference level to multinomial
logisitc regression.}
}
\value{
According to \code{\link{estimateTau}}: a dataframe with two columns,
\code{tauestimate}, \code{varestimate}, where
\code{tauestimate} is a vector of estimates for pairwise treatment effects,
and \code{varestimate} is a vector of variance estimates, using bootstrapping
method.
}
\description{
Stratification on GPS with multilevel treatments
}
\references{
Abadie, A., & Imbens, G. W. (2006). Large sample properties of
  matching estimators for average treatment effects. econometrica, 74(1),
  235-267. \url{https://doi.org/10.1111/j.1468-0262.2006.00655.x}

  Abadie, A., & Imbens, G. W. (2016). Matching on the estimated propensity
  score. Econometrica, 84(2), 781-807.
  \url{https://doi.org/10.3982/ECTA11293}

  Crump, R. K., Hotz, V. J., Imbens, G. W., & Mitnik, O. A. (2009). Dealing
  with limited overlap in estimation of average treatment effects.
  Biometrika, 96(1), 187-199. \url{https://doi.org/10.1093/biomet/asn055}
}
\seealso{
\code{\link{multilevelGPSMatch}}; \code{\link{multilevelMatchX}}
}
