% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelMatching.R
\docType{package}
\name{multilevelMatching}
\alias{multilevelMatching}
\alias{multilevelMatching-package}
\title{Propensity Score Matching and Subclassification in Observational Studies with
Multi-level Treatments}
\description{
\pkg{multilevelMatching} implements the estimators introduced in Yang et al.
(2016) \emph{Propensity Score Matching and Subclassification in Observational
studies with Multi-level Treatments}:
\url{https://doi.org/10.1111/biom.12505}. These are
covariate- and propensity score-matching estimators for estimating the causal
effect of multilevel treatment (i.e., 3 or more treatment types).
}
\details{
For example, see \code{\link{multilevelMatchX}}.
}

