% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevelMatchX.r
\name{multilevelMatchX}
\alias{multilevelMatchX}
\title{Matching on X with multilevel treatments}
\usage{
multilevelMatchX(Y, W, X)
}
\arguments{
\item{Y}{a continuous response vector (1 x n)}

\item{W}{a treatment vector (1 x n) with numerical values indicating treatment groups}

\item{X}{a covariate matrix (p x n) with no intercept}
}
\value{
A list with 2 elements: tauestimate, varestimate, where
tauestimate is a vector of estimates for pairwise treatment effects,
varestimate is a vector of variance estimates for tauestimate, using Abadie&Imbens(2006)'s method.
}
\description{
Matching on X with multilevel treatments
}
\examples{
  X<-c(5.5,10.6,3.1,8.7,5.1,10.2,9.8,4.4,4.9)
  Y<-c(102,105,120,130,100,80,94,108,96)
  W<-c(1,1,1,3,2,3,2,1,2)
  multilevelMatchX(Y,W,X)

}
\seealso{
\code{\link{multilevelGPSMatch}}; \code{\link{multilevelGPSStratification}}
}

