% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameCols}
\alias{nameCols}
\title{Name Two or All Columns in the Matching Matrix}
\usage{
nameCols(trt_levels)
}
\arguments{
\item{trt_levels}{A vector (of length \code{num_trts} providing the unique
treatment levels}
}
\description{
Useful when using Abadie and Imbens (2016) variance estimator
}
