% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameContrast}
\alias{nameContrast}
\title{Naming the matching contrasts}
\usage{
nameContrast(trt1, trt2)
}
\arguments{
\item{trt1}{Former treatment level}

\item{trt2}{Latter treatment level}
}
\description{
Naming the matching contrasts
}
\examples{
  nameContrast(trt1=1, trt2=0)

}
