% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.r
\name{overlap}
\alias{overlap}
\title{Function to determine overlap from Crump et al. (2009)'s method.}
\usage{
overlap(PF.fit)
}
\arguments{
\item{PF.fit}{fitted propensity model}
}
\description{
Function to determine overlap from Crump et al. (2009)'s method.
}
\references{
Crump, R. K., Hotz, V. J., Imbens, G. W., & Mitnik, O. A. (2009).
  Dealing with limited overlap in estimation of average treatment effects.
  Biometrika, 96(1), 187-199. \url{https://doi.org/10.1093/biomet/asn055}
}
