% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidying_utilities.R
\name{reorderByTreatment}
\alias{reorderByTreatment}
\title{order the treatment increasingly}
\usage{
reorderByTreatment(W, X, Y)
}
\arguments{
\item{W}{a treatment vector (1 x n) with numerical values indicating treatment groups}

\item{X}{a covariate matrix (p x n) with no intercept}

\item{Y}{a continuous response vector (1 x n)}
}
\value{
the following three elements, ordered according to levels of W
\itemize{

 \item W: a treatment vector (1 x n) with numerical values indicating treatment groups

 \item X: a covariate matrix (p x n) with no intercept

 \item Y: a continuous response vector (1 x n)
}
}
\description{
order the treatment increasingly
}
