% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidying_utilities.R
\name{reorderByTreatment}
\alias{reorderByTreatment}
\title{order the treatment increasingly}
\usage{
reorderByTreatment(Y, W, X)
}
\arguments{
\item{Y}{a continuous response vector (1 x n)}

\item{W}{a treatment vector (1 x n) with numerical values indicating treatment groups}

\item{X}{a covariate matrix (p x n) with no intercept}
}
\value{
the following elements, ordered according to levels of W
\itemize{

 \item W: a treatment vector (1 x n) with numerical values indicating treatment groups

 \item X: a covariate matrix (p x n) with no intercept

 \item Y: a continuous response vector (1 x n)
}
along with these downstream elements of treatment:
\itemize{
\item trtnumber: number of treatment levels
\item trtlevels: all treatment levels
\item pertrtlevelnumber: number of observations by treatment level
\item taunumber: number of pairwise treatment effects
}
}
\description{
order the treatment increasingly
}
