% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reorderByTreatment}
\alias{reorderByTreatment}
\title{order the treatment increasingly}
\usage{
reorderByTreatment(Y, W, X, unit_ids_unsorted)
}
\arguments{
\item{Y}{a continuous response vector (1 x n)}

\item{W}{a treatment vector (1 x n) with numerical values indicating treatment groups}

\item{X}{a covariate matrix (p x n) with no intercept}

\item{unit_ids_unsorted}{The \code{unit_ids} before the data is reordered}
}
\value{
the following elements, ordered according to levels of \code{W}
\itemize{

 \item \code{W}: a treatment vector (1 x n) with numerical values indicating treatment groups

 \item \code{X}: a covariate matrix (p x n) with no intercept

 \item \code{Y}: a continuous response vector (1 x n)
}
along with these downstream elements of treatment:
\itemize{
\item \code{num_trts}: number of treatment levels
\item \code{trt_levels}: all treatment levels
\item \code{N_per_trt}: number of observations by treatment level
\item \code{num_contrasts}: number of pairwise treatment effects
\item \code{orig_to_sorted}: vector to rearrange from original to sorted by treatment
\item \code{sorted_to_orig}: vector to rearrange from sorted to original order
}
}
\description{
order the treatment increasingly
}
