% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{setIDs}
\alias{setIDs}
\title{Set identifiers from vector/matrix/dataframe}
\usage{
setIDs(x, unit_ids)
}
\arguments{
\item{x}{An object}

\item{unit_ids}{Character vector to identify study units.}
}
\value{
The object \code{x} after setting its \code{names(x)},
  \code{row.names(x)}, or \code{rownames(x)} to \code{unit_ids}.
}
\description{
This is a helper function for \code{\link{determineIDs}}
}

